Splitting up the modules
