+++ atom/src/lib.rs
@@ -4,24 +4,25 @@ pub extern crate lv2_atom_sys as sys;
 extern crate lv2_core as core;
 extern crate lv2_urid as urid;
 
-mod scalar;
-mod slice;
+pub mod scalar;
+pub mod vector;
+pub mod chunk;
+pub mod literal;
 
 pub mod space;
-pub use scalar::*;
-pub use slice::*;
 
 use urid::{URIDCache, URID};
 
 #[derive(URIDCache)]
+/// Container for the URIDs of all `UriBound`s in this crate.
 pub struct AtomURIDCache {
-    pub double: URID<Double>,
-    pub float: URID<Float>,
-    pub int: URID<Int>,
-    pub long: URID<Long>,
-    pub urid: URID<AtomURID>,
-    pub vector: URID<Vector>,
-    pub bool: URID<Bool>,
-    pub chunk: URID<Chunk>,
-    pub literal: URID<Literal>,
+    pub double: URID<scalar::Double>,
+    pub float: URID<scalar::Float>,
+    pub int: URID<scalar::Int>,
+    pub long: URID<scalar::Long>,
+    pub urid: URID<scalar::AtomURID>,
+    pub bool: URID<scalar::Bool>,
+    pub vector: URID<vector::Vector>,
+    pub chunk: URID<chunk::Chunk>,
+    pub literal: URID<literal::Literal>,
 }
